api/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn==0.34.0

# Database (shared with dagster_pipeline)
sqlmodel==0.0.22
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
python-dotenv==1.0.1

# LangChain (semantic search + RAG)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0

# Embeddings
sentence-transformers>=3.0.0